RF-1025: Remove unneeded states
