+++ screens/index.ts
@@ -4,7 +4,7 @@ export { default as AccountsSetupScreen } from './AccountsScreen/AccountsSetupSc
 export { default as ApplicationUnapprovedScreen } from './ApplicationUnapprovedScreen';
 export { default as BankingDisclosuresScreen } from './BankingDisclosuresScreen';
 export { default as ConfirmPIIScreen } from './ConfirmPIIScreen';
-export { default as DisclosuresScreen } from './DisclosuresScreen';
+export { default as DisclosuresScreen } from './Disclosures/DisclosuresScreen';
 export { default as ExternalAccountScreen } from './ExternalAccountScreen';
 export { default as ForgotPasswordScreen } from './ForgotPasswordScreen/ForgotPasswordScreen';
 export { default as InitTransferScreen } from './InitTransferScreen';
@@ -12,7 +12,6 @@ export { default as LoginScreen } from './LoginScreen/LoginScreen';
 export { default as MenuScreen } from './MenuScreen';
 export { default as PatriotActScreen } from './PatriotActScreen';
 export { default as PDFReaderScreen } from './PDFReaderScreen';
-export { default as PIIScreen } from './PIIScreen';
 export { default as ProcessingApplicationScreen } from './ProcessingApplicationScreen';
 export { default as SetPasswordScreen } from './SetPasswordScreen';
 export { default as SignupScreen } from './SignupScreen';
